requirements.txt
pymodbus~=2.5.2
pycryptodome~=3.10.1
PyYAML~=5.4.1
